requirements.txt
openai==0.27.7
newspaper3k==0.2.8 
beautifulsoup4==4.12.2 
urllib3==1.26.15 
requests==2.30.0 
bs4==0.0.1